fix: rename internal
